move proto_utils to private
