+++ src/viam/sdk/services/navigation.cpp
@@ -1,7 +1,7 @@
 #include <viam/sdk/services/navigation.hpp>
 
 #include <viam/api/service/navigation/v1/navigation.pb.h>
-#include <viam/sdk/common/proto_utils.hpp>
+#include <viam/sdk/common/private/proto_utils.hpp>
 #include <viam/sdk/common/utils.hpp>
 
 namespace viam {
