+++ src/viam/sdk/services/private/navigation_server.cpp
@@ -3,7 +3,7 @@
 #include <grpcpp/support/status.h>
 
 #include <viam/sdk/common/pose.hpp>
-#include <viam/sdk/common/proto_utils.hpp>
+#include <viam/sdk/common/private/proto_utils.hpp>
 #include <viam/sdk/common/proto_value.hpp>
 #include <viam/sdk/common/service_helper.hpp>
 #include <viam/sdk/common/utils.hpp>
